Filter questions and answers from Summary that have conditions that are not met
