+++ apps/melding-form/src/app/(general)/aanvullende-vragen/[classificationId]/[panelId]/page.tsx
@@ -14,12 +14,8 @@ import type {
 
 import { getFormClassificationByClassificationId, getMeldingByMeldingIdAnswersMelder } from '@meldingen/api-client'
 
-import {
-  AFTER_ADDITIONAL_QUESTIONS_PATH,
-  getPreviousAnswersByKey,
-  getPreviousPanelPath,
-  isPanelComponentOutput,
-} from './_utils/navigationUtils'
+import { getAnswersByKey, isPanelComponentOutput } from '../../../_utils/getAnswersByKey'
+import { AFTER_ADDITIONAL_QUESTIONS_PATH, getPreviousPanelPath } from './_utils/navigationUtils'
 import { setDateComponentOptions } from './_utils/setDateComponentOptions'
 import { postForm } from './actions'
 import { AdditionalQuestions } from './AdditionalQuestions'
@@ -154,7 +150,7 @@ export default async ({ params }: { params: Params }) => {
     key,
   }))
 
-  const previousAnswersByKey = getPreviousAnswersByKey(data, answers)
+  const previousAnswersByKey = getAnswersByKey(data, answers)
 
   const extraArgs = {
     classificationId,
